fix /sys/devices/system/cpu/online in cpuset not correct bug
When the container runs in cpuset and cpu quota, for example, cpu quota
is limited to 8 cores, cpuset is limited to 1-3, 4-5, then there will be
problems. The core number obtained by /proc/cpuinfo is 4 cores, and
/sys/devices/system/cpu/online gets 8 cores.

The reason is that /sys/devices/system/cpu/online does not take into
account the situation of cpuset. So submit this patch.

Test case:
  docker run -it -d --cpuset-cpus="1-2,3-4" \
    --cpu-period=100000 \
    --cpu-quota=800000 \
    --memory 2048m \
    -v /root/tce/lxcfs/run/lxcfs/sys/devices/system/cpu/online:/sys/devices/system/cpu/online \
    -v /root/tce/lxcfs/run/lxcfs/proc/cpuinfo:/proc/cpuinfo:rw \
    debian:latest bash

Signed-off-by: Hongbo Yin <yinhongbo@bytedance.com>
Signed-off-by: river <river@vvl.me>